+++ src/main/java/no/vegvesen/nvdbapi/client/model/roadnet/roadsysref/RoadCategory.java
@@ -4,16 +4,18 @@ package no.vegvesen.nvdbapi.client.model.roadnet.roadsysref;
  * See definition of road object type 915, attribute 11276, «Vegkategori»
  */
 public enum RoadCategory {
-    E( "Europaveg"),
-    R( "Riksveg"),
-    F( "Fylkesveg"),
-    K( "Kommunal veg"),
-    P( "Privat veg"),
-    S("Skogsveg");
+    E( "Europaveg", "europaveg"),
+    R( "Riksveg", "riksveg"),
+    F( "Fylkesveg", "fylkesveg"),
+    K( "Kommunal veg", "kommunalVeg"),
+    P( "Privat veg", "privatVeg"),
+    S("Skogsveg", "skogsVeg");
 
     public final String name;
+    public final String codeName;
 
-    RoadCategory(String name) {
+    RoadCategory(String name, String codeName) {
         this.name = name;
+        this.codeName = codeName;
     }
 }